feat(new-package): add jest-prosemirror testing library
