+++ base.tsconfig.json
@@ -19,6 +19,7 @@
     "alwaysStrict": true,
     "resolveJsonModule": true,
     "baseUrl": "./",
+    "plugins": [{ "name": "typescript-styled-plugin" }, { "name": "typescript-tslint-plugin" }],
     "paths": {
       "@remirror/core": ["@remirror/core/src/index.ts"],
       "@remirror/core-extensions": ["@remirror/core-extensions/src/index.ts"],
@@ -31,7 +32,7 @@
       "@remirror/renderer-react": ["@remirror/renderer-react/src/index.ts"],
       "@remirror/ui-twitter": ["@remirror/ui-twitter/src"],
       "jest-remirror": ["packages/jest-remirror/src"],
-      "@test-utils": ["@remirror/core/src/__tests__/test-utils.tsx"],
+      "jest-prosemirror": ["packages/jest-prosemirror/src"],
       "@test-fixtures/*": ["support/fixtures/*"],
       "*": ["support/types/*"]
     }
