+++ docs/tsconfig.json
@@ -5,7 +5,6 @@
     "jsx": "react",
     "outDir": "lib",
     "declaration": true,
-    "plugins": [{ "name": "typescript-styled-plugin" }],
     "types": ["jest", "jest-environment-puppeteer", "expect-puppeteer", "puppeteer", "jest-extended"]
   },
   "include": ["./", "patches.d.ts"]
