+++ packages/jest-prosemirror/.babelrc.js
@@ -1,0 +1,1 @@
+module.exports = require('../../support/babel/base.babel');
