+++ packages/jest-prosemirror/environment.js
@@ -1,0 +1,1 @@
+require('./lib/environment').setupEnvironment();
