+++ packages/jest-prosemirror/jest.config.js
@@ -1,0 +1,7 @@
+const config = require('../../support/jest/jest.config');
+
+module.exports = {
+  ...config,
+  name: 'jest-prosemirror',
+  displayName: 'jest-prosemirror',
+};
