+++ packages/jest-prosemirror/readme.md
@@ -1,0 +1,27 @@
+<div align="center">
+	<br />
+	<div>
+		<img width="300" src="../../support/assets/logo-icon.svg" alt="remirror" />
+    <h1 align="center">jest-prosemirror</h1>
+	</div>
+    <br />
+    <br />
+    <br />
+    <br />
+</div>
+
+[![npm](https://img.shields.io/npm/dm/jest-prosemirror.svg?style=for-the-badge&logo=npm)](https://www.npmjs.com/package/jest-prosemirror) [![Dependencies (path)](https://img.shields.io/david/ifiokjr/remirror.svg?logo=npm&path=@remirror%2Fextension-mention&style=for-the-badge)](https://github.com/ifiokjr/remirror/blob/master/jest-prosemirror/package.json) [![NPM](https://img.shields.io/npm/l/jest-prosemirror.svg?style=for-the-badge)](https://github.com/ifiokjr/remirror/blob/master/LICENSE) [![GitHub issues by-label](https://img.shields.io/github/issues/ifiokjr/remirror/package%3A%20jest-prosemirror.svg?label=Open%20Issues&logo=github&style=for-the-badge)](https://github.com/ifiokjr/remirror/issues?utf8=%E2%9C%93&q=is%3Aissue+is%3Aopen+sort%3Aupdated-desc+label%3Apackage%3A%20jest-prosemirror) [![GitHub pull requests by-label](https://img.shields.io/github/issues-pr/ifiokjr/remirror/package%3A%20jest-prosemirror.svg?label=Open%20Pull%20Requests&logo=github&style=for-the-badge)](https://github.com/ifiokjr/remirror/pulls?utf8=%E2%9C%93&q=is%3Apr+is%3Aopen+sort%3Aupdated-desc+label%3Apackage%3A%20jest-prosemirror)
+
+## Installation
+
+```bash
+yarn add jest-prosemirror
+```
+
+## Getting started
+
+API and docs still being worked on.
+
+## Acknowledgements
+
+This package borrows very heavily from [@atlaskit/editor-test-helpers](https://www.npmjs.com/package/@atlaskit/editor-test-helpers)
