+++ packages/jest-prosemirror/src/environment.ts
@@ -1,0 +1,8 @@
+import { prosemirrorMatchers } from './matchers';
+
+/**
+ * Setup the environment automatically for jest-prosemirror
+ */
+export const setupEnvironment = () => {
+  expect.extend(prosemirrorMatchers);
+};
