+++ packages/jest-prosemirror/src/index.ts
@@ -1,0 +1,3 @@
+export * from './matchers';
+export * from './utils';
+export * from './environment';
