+++ packages/jest-prosemirror/tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "extends": "../../base.tsconfig.json",
+  "compilerOptions": {
+    "lib": ["esnext", "dom"],
+    "jsx": "react",
+    "outDir": "lib",
+    "types": ["jest", "jest-extended", "jest-axe", "jest-dom/extend-expect"]
+  },
+  "include": ["src"]
+}
