+++ packages/jest-prosemirror/tsconfig.prod.json
@@ -1,0 +1,12 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": { "baseUrl": "src", "paths": { "*": ["../../../support/types/*"] } },
+  "exclude": [
+    "**/*.test.{ts,tsx}",
+    "**/*.stories.{ts,tsx}",
+    "**/*.spec.{ts,tsx}",
+    "**/__mocks__/**",
+    "**/__tests__/**",
+    "**/__stories__/**"
+  ]
+}
