+++ packages/jest-remirror/src/setup-environment.ts
@@ -1,21 +1,16 @@
 import { environment } from '@remirror/core';
-import { toHaveNoViolations } from 'jest-axe';
-import { createSerializer, matchers } from 'jest-emotion';
+
 import { jsdomExtras, jsdomPolyfill } from './jsdom-polyfills';
 import { remirrorMatchers } from './matchers';
 
 export const setupRemirrorEnvironment = () => {
-  // The following can only be run in a DOM environment
+  /* Only continue when in a DOM environment */
   if (!environment.isBrowser) {
     return;
   }
 
+  /* Auto cleanup DOM after each test */
   require('react-testing-library/cleanup-after-each');
-  require('jest-dom/extend-expect');
-  expect.addSnapshotSerializer(createSerializer({}));
-
-  expect.extend(toHaveNoViolations);
-  expect.extend(matchers);
 
   /* Add matchers to jest */
   expect.extend(remirrorMatchers);
