+++ support/jest/jest.config.js
@@ -39,9 +39,9 @@ module.exports = {
   cacheDirectory: baseDir('.jest'),
   testEnvironment: 'jsdom',
   moduleNameMapper: {
-    '@test-utils$': baseDir('@remirror', 'core', 'src', '__tests__', 'test-utils.tsx'),
     '@remirror\\/([a-z0-9-]+)$': baseDir('@remirror/$1/src'),
     'jest-remirror$': baseDir('packages', 'jest-remirror', 'src'),
+    'jest-prosemirror$': baseDir('packages', 'jest-prosemirror', 'src'),
     '@test-fixtures/(.*)$': baseDir('support/fixtures/$1'),
   },
   testRunner: 'jest-circus/runner',
